@@ -5,14 +5,19 @@ import android.database.Cursor;
 import android.database.SQLException;
 import android.database.sqlite.SQLiteDatabase;
 import android.database.sqlite.SQLiteException;
-import android.support.v7.app.AppCompatActivity;
+//import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.AdapterView;
 import android.widget.ListView;
+
+import androidx.appcompat.app.AppCompatActivity;
+
 import java.util.ArrayList;
 
-public class MainActivity extends AppCompatActivity implements DownloadEqsAsyncTask.DownloadEqsInterface{
+import richard.terremotosenelmundo.DownloadEqsAsyncTask.DownloadEqsInterface;
+
+public class MainActivity extends AppCompatActivity implements DownloadEqsInterface{
 
     public final static String SELECTED_EARTHQUAKE = "selected_earthquake";
     private ListView earthquakeListView;
@@ -23,7 +28,7 @@ public class MainActivity extends AppCompatActivity implements DownloadEqsAsyncT
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
 
-        earthquakeListView = (ListView) findViewById(R.id.earthquake_list_view);
+        earthquakeListView = findViewById(R.id.earthquake_list_view);
 
         if(Utils.isNetworkAvailable(this)){
 
